Factor out payment processor views into seperate view files.
