+++ ecommerce/extensions/payment/urls.py
@@ -1,11 +1,12 @@
 """ Payment-related URLs """
 from django.conf.urls import url
 
-from ecommerce.extensions.payment import views
+from ecommerce.extensions.payment.views.cybersource import CybersourceNotifyView, CybersourceSubmitView
+from ecommerce.extensions.payment.views.paypal import PaypalPaymentExecutionView, PaypalProfileAdminView
 
 urlpatterns = [
-    url(r'^cybersource/notify/$', views.CybersourceNotifyView.as_view(), name='cybersource_notify'),
-    url(r'^cybersource/submit/$', views.CybersourceSubmitView.as_view(), name='cybersource_submit'),
-    url(r'^paypal/execute/$', views.PaypalPaymentExecutionView.as_view(), name='paypal_execute'),
-    url(r'^paypal/profiles/$', views.PaypalProfileAdminView.as_view(), name='paypal_profiles'),
+    url(r'^cybersource/notify/$', CybersourceNotifyView.as_view(), name='cybersource_notify'),
+    url(r'^cybersource/submit/$', CybersourceSubmitView.as_view(), name='cybersource_submit'),
+    url(r'^paypal/execute/$', PaypalPaymentExecutionView.as_view(), name='paypal_execute'),
+    url(r'^paypal/profiles/$', PaypalProfileAdminView.as_view(), name='paypal_profiles'),
 ]
